Created BMI calculator
this code receives input from the user of a person's age and height and then calculates the person's BMI using the 2 inputs using the BMI formula of dividing weight by the height squared.
it then outputs the user's BMI with the category it falls under. i.e under weight, obese, Normal