Created a count down timer
imported a python function called time.
user inputs time in seconds, and the code uses a for loop set the timer's count down in reference to the user's input  and also using  the sleep modules from the time function.
the time counts down to the last second and then the loop stops and the user is told time is up.